+++ NEWS.md
@@ -1,3 +1,10 @@
+# PlaneGeometry 1.6.0 (2023-08-07)
+
+- New functions `LineFromInterceptAndSlope` and `LineFromEquation`.
+
+- New R6 class `Hyperbola`.
+
+
 # PlaneGeometry 1.5.1 (2023-06-23)
 
 - Unit tests and vignette chunks are now skipped if dependencies are missing.